+++ YoutubeKeysFix.user.js
@@ -257,7 +257,14 @@
         s.type= 'text/css';
         s.textContent= `
 
-#player-container:focus-within { box-shadow: 0 0 20px 0px rgba(0,0,0,0.8); }
+/* Add permanent box-shadow under player */
+/* #player-full-bleed-container */
+#player-container {
+  box-shadow: 0 0.5em 3em 0px rgba(0,0,0,0.7);
+}
+ytd-channel-video-player-renderer {
+  overflow: visible;
+}
 
 /* Highlight focused button in player */
 .ytp-probably-keyboard-focus :focus {